pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "is3d-native"
version = "0.1.0"
description = "Windows-native training scaffold for is3d (FastViT + TriPlane)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "torch>=2.3",
  "torchvision>=0.18",
  "webdataset>=0.2.96",
  "numpy>=1.26",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.5",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
